BETA: Fix for numeric type from #4682 by mrTJO, thanks Lineage for report.
git-svn-id: http://svn.l2jserver.com/branches/unstable@4720 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/CharSelectInfoPackage.java
@@ -56,7 +56,7 @@ public class CharSelectInfoPackage
 	private int _z = 0;
 	private String _htmlPrefix = null;
 	private int _vitalityPoints = 0;
-	private byte _accessLevel = 0;
+	private int _accessLevel = 0;
 	
 	/**
 	 * Constructor for CharSelectInfoPackage.
@@ -86,7 +86,7 @@ public class CharSelectInfoPackage
 	/**
 	 * @return the character's access level.
 	 */
-	public byte getAccessLevel()
+	public int getAccessLevel()
 	{
 		return _accessLevel;
 	}
@@ -94,7 +94,7 @@ public class CharSelectInfoPackage
 	/**
 	 * @param level the character's access level to be set.
 	 */
-	public void setAccessLevel(byte level)
+	public void setAccessLevel(int level)
 	{
 		_accessLevel = level;
 	}
